==> 04_split_fitting_parameters.ipynb <==
# 04 Splitting the Fitting Parameters

Some of the fitting parameters are constant across all experiments, so they do not need to be predicted. Here, the fitting parameters are divided into those that are constant and those which need to be predicted.

import pandas as pd

df = pd.read_csv('./training_data/03_fit_params_standard.csv')

params_to_predict_min = []
params_to_predict_max = []
params_to_predict_names = []
params_constant_values = []
params_constant_names = []
for fit_param_name in df.columns:
    param = df[fit_param_name]
    param_min = min(param)
    param_max = max(param)
    if param_min == param_max:
        params_constant_names.append(fit_param_name)
        params_constant_values.append(param_min)
    else:
        params_to_predict_names.append(fit_param_name)
        params_to_predict_min.append(param_min)
        params_to_predict_max.append(param_max)

# Write the min and max values to a file
filename = './training_data/04_fit_params_to_predict.csv'
with open(filename, 'w') as f:
    #write header
    for i in range(len(params_to_predict_names)):
        f.write(params_to_predict_names[i])
        if i < len(params_to_predict_names) - 1:
            f.write(',')
    f.write('\n')
    for i in range(len(params_to_predict_min)):
        f.write(str(params_to_predict_min[i]))
        if i < len(params_to_predict_min) - 1:
            f.write(',')
    f.write('\n')
    for i in range(len(params_to_predict_max)):
        f.write(str(params_to_predict_max[i]))
        if i < len(params_to_predict_max) - 1:
            f.write(',')

# Write the constant values to a file
filename = './training_data/04_fit_params_constant.csv'
with open(filename, 'w') as f:
    #write header
    for i in range(len(params_constant_names)):
        f.write(params_constant_names[i])
        if i < len(params_constant_names) - 1:
            f.write(',')
    f.write('\n')
    for i in range(len(params_constant_values)):
        f.write(str(params_constant_values[i]))
        if i < len(params_constant_values) - 1:
            f.write(',')
