==> 13_train_full_models.ipynb <==
## 13 Train Full Models
After tuning the hyperparameters with 5-fold cross validation, the models are all trained on the entire training dataset to expose them to the most data.

import pandas as pd
import numpy as np
from sklearn.ensemble import RandomForestRegressor
from sklearn.neural_network import MLPRegressor
import xgboost as xgb
import pickle

spectra_normalized = pd.read_csv('./training_data/08_train_int_spectra.csv', header=None)
fit_params_standard = pd.read_csv('./training_data/07_train_fit_params_standard.csv')
fit_params_constant = pd.read_csv('./training_data/04_fit_params_constant.csv')

print(len(spectra_normalized))
print(len(fit_params_standard))

x_train = spectra_normalized
y_train = fit_params_standard.drop(columns=fit_params_constant.columns)

xgb = xgb.XGBRegressor(eta=0.2, gamma=0, alpha=0, reg_lambda=0, max_depth=7)
rf = RandomForestRegressor(n_estimators=500, min_samples_leaf=1, max_features="sqrt", min_samples_split=2)
mlp = MLPRegressor(early_stopping=True, batch_size=10, alpha=0.001, learning_rate="constant", 
                   learning_rate_init=0.001, max_iter=500, hidden_layer_sizes=(25, 50, 75, 100))

xgb.fit(x_train, y_train)

rf.fit(x_train, y_train)

mlp.fit(x_train, y_train)

pickle.dump(rf, open('13_random_forest.sav', 'wb'))
pickle.dump(mlp, open('13_multilayer_perceptron.sav', 'wb'))
pickle.dump(xgb, open('13_xgboost.sav', 'wb'))
