==> 17_timing.ipynb <==
## 17 Timing
Fitting parameters should be provided quicker by the models than a scientist by hand, so here the models are timed on a handful of predictions.

import pandas as pd
import pickle
import xgboost

spectra = pd.read_csv('./training_data/08_test_int_spectra.csv', header=None)

%timeit xgb_model = pickle.load(open('13_xgboost.sav', 'rb'))

xgb_model = pickle.load(open('13_xgboost.sav', 'rb'))
%timeit xgb_model.predict(spectra.iloc[0:4])

%timeit mlp_model = pickle.load(open('13_multilayer_perceptron.sav', 'rb'))

mlp_model = pickle.load(open('13_multilayer_perceptron.sav', 'rb'))
%timeit mlp_model.predict(spectra.iloc[0:4])

%timeit rf_model = pickle.load(open('13_random_forest.sav', 'rb'))

rf_model = pickle.load(open('13_random_forest.sav', 'rb'))
%timeit rf_model.predict(spectra.iloc[0:4])
